Add paddlespeech.resource.
